Lazy load components
Change-type: minor
