+++ apps/ui/lib/components/MarkdownEditor.tsx
@@ -1,0 +1,2 @@
+import { MarkdownWidget } from 'rendition/dist/extra/Form/markdown';
+export default MarkdownWidget;
