+++ apps/ui/lib/components/SafeLazy.tsx
@@ -5,15 +5,20 @@ import Splash from './Splash';
 const SafeLazy = ({ children }) => {
 	return (
 		<ErrorBoundary>
-			<Suspense fallback={<Splash />}>{children}</Suspense>
+			<Suspense fallback={<Splash data-testid="splash" />}>{children}</Suspense>
 		</ErrorBoundary>
 	);
 };
 
-export const createLazyComponent = (fn) => {
+export const createLazyComponent = <TProps extends any>(
+	fn: () => Promise<{ default: React.ComponentType<TProps> }>,
+) => {
 	const LazyComponent = React.lazy(fn);
 
-	return (props) => {
+	return (
+		props: JSX.IntrinsicAttributes &
+			React.PropsWithRef<TProps & { children?: React.ReactNode }>,
+	) => {
 		return (
 			<SafeLazy>
 				<LazyComponent {...props} />
