+++ apps/ui/lib/components/Widgets/MarkdownWidget.tsx
@@ -1,0 +1,2 @@
+import { MarkdownWidget } from 'rendition/dist/extra/Renderer/MarkdownWidget';
+export default MarkdownWidget;
