+++ apps/ui/lib/components/Widgets/MermaidWidget.tsx
@@ -1,0 +1,2 @@
+import { MermaidWidget } from 'rendition/dist/extra/Renderer/MermaidWidget';
+export default MermaidWidget;
