+++ apps/ui/lib/components/Widgets/index.ts
@@ -1,10 +1,16 @@
-import { MarkdownWidget } from 'rendition/dist/extra/Renderer/MarkdownWidget';
-import { MermaidWidget } from 'rendition/dist/extra/Renderer/MermaidWidget';
 import { withOptionProps } from 'rendition/dist/components/Renderer/widgets/widget-util';
 import { JellyfishLinkWidget } from './JellyfishLinkWidget';
 import { JellyfishUserWidget } from './JellyfishUserWidget';
-
+import { createLazyComponent } from '../SafeLazy';
 export { LoopSelectWidget } from './LoopSelectWidget';
+
+export const MarkdownWidget = createLazyComponent(
+	() => import(/* webpackChunkName: "markdown-widget" */ './MarkdownWidget'),
+);
+
+export const MermaidWidget = createLazyComponent(
+	() => import(/* webpackChunkName: "mermaid-widget" */ './MermaidWidget'),
+);
 
 export const JellyfishWidgets = [
 	{
