+++ apps/ui/lib/lens/actions/CreateUserLens/index.tsx
@@ -1,8 +1,12 @@
 import _ from 'lodash';
 import { connect } from 'react-redux';
 import * as redux from 'redux';
+import { createLazyComponent } from '../../../components/SafeLazy';
 import { actionCreators, selectors } from '../../../core';
-import CreateUserLens from './CreateUserLens';
+
+export const CreateUserLens = createLazyComponent(
+	() => import(/* webpackChunkName: "lens-create-user" */ './CreateUserLens'),
+);
 
 const mapStateToProps = (state, ownProps) => {
 	return {
