+++ apps/ui/lib/lens/full/CheckRun/index.tsx
@@ -1,8 +1,12 @@
 import _ from 'lodash';
 import { connect } from 'react-redux';
 import { bindActionCreators } from 'redux';
+import { createLazyComponent } from '../../../components/SafeLazy';
 import { actionCreators, selectors } from '../../../core';
-import CheckRun from './CheckRun';
+
+export const CheckRun = createLazyComponent(
+	() => import(/* webpackChunkName: "lens-check-run" */ './CheckRun'),
+);
 
 const mapStateToProps = (state) => {
 	return {
