+++ apps/ui/lib/lens/full/Thread/Thread.tsx
@@ -1,15 +1,13 @@
 import { circularDeepEqual } from 'fast-equals';
 import _ from 'lodash';
 import React from 'react';
-import { connect } from 'react-redux';
 import { Box, Txt } from 'rendition';
 import { Collapsible, helpers, TagList } from '@balena/jellyfish-ui-components';
-import CardFields from '../../components/CardFields';
-import { selectors } from '../../core';
-import Timeline from '../list/Timeline';
-import CardLayout from '../../layouts/CardLayout';
+import CardFields from '../../../components/CardFields';
+import Timeline from '../../list/Timeline';
+import CardLayout from '../../../layouts/CardLayout';
 
-class Thread extends React.Component<any, any> {
+export default class Thread extends React.Component<any, any> {
 	shouldComponentUpdate(nextProps) {
 		return !circularDeepEqual(nextProps, this.props);
 	}
@@ -60,32 +58,3 @@ class Thread extends React.Component<any, any> {
 		);
 	}
 }
-
-const mapStateToProps = (state) => {
-	return {
-		types: selectors.getTypes(state),
-	};
-};
-
-const lens = {
-	slug: 'lens-default',
-	type: 'lens',
-	version: '1.0.0',
-	name: 'Default lens',
-	data: {
-		icon: 'address-card',
-		format: 'full',
-		renderer: connect(mapStateToProps)(Thread),
-		filter: {
-			type: 'object',
-			properties: {
-				type: {
-					type: 'string',
-					const: 'thread@1.0.0',
-				},
-			},
-		},
-	},
-};
-
-export default lens;
