+++ apps/ui/lib/lens/full/Thread/index.ts
@@ -1,0 +1,36 @@
+import { connect } from 'react-redux';
+import { createLazyComponent } from '../../../components/SafeLazy';
+import { selectors } from '../../../core';
+
+export const Thread = createLazyComponent(
+	() => import(/* webpackChunkName: "lens-thread" */ './Thread'),
+);
+
+const mapStateToProps = (state) => {
+	return {
+		types: selectors.getTypes(state),
+	};
+};
+
+const lens = {
+	slug: 'lens-default',
+	type: 'lens',
+	version: '1.0.0',
+	name: 'Default lens',
+	data: {
+		icon: 'address-card',
+		format: 'full',
+		renderer: connect(mapStateToProps)(Thread),
+		filter: {
+			type: 'object',
+			properties: {
+				type: {
+					type: 'string',
+					const: 'thread@1.0.0',
+				},
+			},
+		},
+	},
+};
+
+export default lens;
