+++ apps/ui/lib/lens/snippet/SingleCard/SingleContract.tsx
