+++ apps/ui/package.json
@@ -92,6 +92,7 @@
     "@balena/lint": "^6.2.0",
     "@emotion/cache": "^11.7.1",
     "@emotion/react": "^11.7.1",
+    "@testing-library/react": "^12.1.2",
     "@types/common-tags": "^1.8.1",
     "@types/debounce-promise": "^3.1.4",
     "@types/enzyme": "^3.10.11",
